+++ CHANGELOG.md
@@ -26,6 +26,7 @@ hangeLog
 - BUGFIX: fix "t2t-sentence" node by applying defensive guards for out-of-bound touches in queue
 - BUGFIX: fix "t2t-sentence" node by avoiding listener accumulation and duplicate chunk pushes
 - BUGFIX: fix "t2t-sentence" node by ensuring its workOffQueue does not drop calls on re-registration
+- IMPROVEMENT: improve "t2t-sentence" node by using a more elaborative punctuation matching to support more edge cases
 - REFACTOR: split work off queue in "t2t-sentence" into outer and inner parts
 - CLEANUP: cleanup result type of audio classification pipeline in gender node